+++ src/domains/lightning/usecases/service/lightning_payments.rs
@@ -1,150 +1,66 @@
 use async_trait::async_trait;
-use breez_sdk_core::{Payment, PaymentDetails, PaymentFailedData, PaymentStatus, PaymentType};
-use tracing::{info, trace};
+use tracing::{debug, info};
 use uuid::Uuid;
 
 use crate::{
     application::errors::{ApplicationError, DataError},
-    domains::lightning::{
-        adapters::LightningRepository,
-        entities::{
-            LightningInvoice, LightningInvoiceStatus, LightningPayment, LightningPaymentStatus,
-        },
-        usecases::LightningPaymentsUseCases,
+    domains::{
+        lightning::{entities::LightningPayment, usecases::LightningPaymentsUseCases},
+        users::entities::{AuthUser, Permission},
     },
 };
 
-pub struct LightningPaymentsProcessor {
-    pub store: Box<dyn LightningRepository>,
-}
-
-impl LightningPaymentsProcessor {
-    pub fn new(store: Box<dyn LightningRepository>) -> Self {
-        LightningPaymentsProcessor { store }
-    }
-}
+use super::LightningService;
 
 #[async_trait]
-impl LightningPaymentsUseCases for LightningPaymentsProcessor {
-    async fn process_incoming_payment(
+impl LightningPaymentsUseCases for LightningService {
+    async fn get_lightning_payment(
         &self,
-        payment: Payment,
-    ) -> Result<LightningInvoice, ApplicationError> {
-        let payment_hash = payment.id;
-        trace!(payment_hash, "Processing incoming lightning payment");
-
-        if payment.status != PaymentStatus::Complete {
-            return Err(DataError::Validation("Payment is not Complete.".into()).into());
-        }
-
-        if payment.payment_type != PaymentType::Received {
-            return Err(DataError::Validation("Payment is not Received.".into()).into());
-        }
-
-        let invoice_option = self.store.find_invoice_by_hash(&payment_hash).await?;
-
-        if let Some(mut invoice) = invoice_option {
-            invoice.fee_msat = Some(payment.fee_msat);
-            invoice.status = LightningInvoiceStatus::SETTLED;
-            invoice.payment_time = Some(payment.payment_time);
-
-            invoice = self.store.update_invoice(invoice).await?;
-
-            info!(
-                payment_hash,
-                "Incoming Lightning payment processed successfully"
-            );
-            return Ok(invoice);
-        }
-
-        return Err(DataError::NotFound("Lightning invoice not found.".into()).into());
-    }
-
-    async fn process_outgoing_payment(
-        &self,
-        payment_success: Payment,
+        user: AuthUser,
+        id: Uuid,
     ) -> Result<LightningPayment, ApplicationError> {
-        let payment_id = match payment_success.details.clone() {
-            PaymentDetails::Ln { data } => {
-                Uuid::parse_str(&data.label).map_err(|e| DataError::Validation(e.to_string()))
-            }
-            _ => {
-                Err(DataError::NotFound("Lightning payment not found. Missing label".into()).into())
-            }
-        }?;
-        let payment_hash = payment_success.id.clone();
-        trace!(
-            payment_id = payment_id.to_string(),
-            payment_hash,
-            "Processing outgoing lightning payment"
-        );
+        debug!(user_id = user.sub, "Fetching lightning payment");
 
-        if payment_success.status != PaymentStatus::Complete {
-            return Err(DataError::Validation("Payment is not Complete.".into()).into());
-        }
-
-        if payment_success.payment_type != PaymentType::Sent {
-            return Err(DataError::Validation("Payment is not Sent.".into()).into());
-        }
-
-        let payment_option = self.store.find_payment(payment_id).await?;
-
-        if let Some(payment_retrieved) = payment_option {
-            // We overwrite the payment with the new one at the correct status
-            let mut payment: LightningPayment = payment_success.clone().into();
-            payment.id = payment_retrieved.id;
-            payment.status = LightningPaymentStatus::SETTLED;
-
-            let payment = self.store.update_payment(payment).await?;
-
-            info!(
-                payment_id = payment_id.to_string(),
-                payment_hash,
-                username = payment.lightning_address,
-                "Outgoing Lightning payment processed successfully"
-            );
-            return Ok(payment);
-        }
-
-        return Err(DataError::NotFound("Lightning payment not found.".into()).into());
-    }
-
-    async fn process_failed_payment(
-        &self,
-        payment_failed: PaymentFailedData,
-    ) -> Result<LightningPayment, ApplicationError> {
-        if payment_failed.invoice.is_none() {
-            return Err(
-                DataError::Validation("Failed payment is missing an Invoice.".into()).into(),
-            );
-        }
-
-        let invoice = payment_failed.invoice.unwrap();
-        trace!(
-            payment_hash = invoice.payment_hash,
-            "Processing outgoing failed lightning payment"
-        );
-
-        let payment_option = self
+        let lightning_payment = self
             .store
-            .find_payment_by_hash(&invoice.payment_hash)
-            .await?;
+            .find_payment(id)
+            .await?
+            .ok_or_else(|| DataError::NotFound("Lightning payment not found.".to_string()))?;
 
-        if let Some(mut payment) = payment_option {
-            payment.status = LightningPaymentStatus::FAILED;
-            payment.payment_time = Some(invoice.timestamp as i64);
-            payment.error = Some(payment_failed.error);
-
-            payment = self.store.update_payment(payment).await?;
-
-            info!(
-                payment_hash = invoice.payment_hash,
-                username = payment.lightning_address,
-                "Outgoing Lightning payment processed successfully"
-            );
-            return Ok(payment);
+        if lightning_payment.user_id != user.sub {
+            user.check_permission(Permission::ReadLightningAccounts)?;
         }
 
-        return Err(DataError::NotFound("Lightning payment not found.".into()).into());
+        info!(
+            user_id = user.sub,
+            id = id.to_string(),
+            "Lightning payment fetched successfully"
+        );
+        Ok(lightning_payment)
+    }
+
+    async fn list_lightning_payments(
+        &self,
+        user: AuthUser,
+        limit: Option<u64>,
+        offset: Option<u64>,
+    ) -> Result<Vec<LightningPayment>, ApplicationError> {
+        debug!(
+            user_id = user.sub,
+            limit, offset, "Listing lightning payments"
+        );
+
+        let lightning_payments = if user.has_permission(Permission::ReadLightningAccounts) {
+            // The user has permission to view all addresses
+            self.store.find_all_payments(None, limit, offset).await?
+        } else {
+            // The user can only view their own payments
+            self.store
+                .find_all_payments(Some(user.sub.clone()), limit, offset)
+                .await?
+        };
+
+        info!(user_id = user.sub, "Lightning payments listed successfully");
+        Ok(lightning_payments)
     }
 }
