+++ src/domains/lightning/usecases/service/mod.rs
@@ -1,7 +1,10 @@
 mod lightning_addresses;
+mod lightning_invoices;
 mod lightning_node;
 mod lightning_payments;
+mod lightning_payments_processor;
 mod lightning_service;
+mod wallet;
 
-pub use lightning_payments::LightningPaymentsProcessor;
+pub use lightning_payments_processor::LightningPaymentsProcessor;
 pub use lightning_service::LightningService;
